requirements.txt
# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# PDF Processing
pypdf2>=3.0.0
pdfplumber>=0.10.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL direct connection

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# HTTP
requests>=2.31.0

# Template
jinja2>=3.1.0

# CLI
typer>=0.12.0
rich>=13.0.0

# Utilities
tiktoken>=0.5.0  # Token counting for OpenAI

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
